refactor :: 지원서 상태 enums 패키지 이동
