+++ src/main/java/com/example/daedongv3_5/domain/recruitment/domain/enums/RecruitmentStatus.java
@@ -1,4 +1,4 @@
-package com.example.daedongv3_5.domain.recruitment.domain;
+package com.example.daedongv3_5.domain.recruitment.domain.enums;
 
 public enum RecruitmentStatus {
     SUBMITTED, // 제출됨
