+++ src/main/java/com/example/daedongv3_5/domain/recruitment/presentation/dto/response/RecruitmentResponse.java
@@ -1,8 +1,7 @@
 package com.example.daedongv3_5.domain.recruitment.presentation.dto.response;
 
-import com.example.daedongv3_5.domain.club.domain.Club;
 import com.example.daedongv3_5.domain.club.domain.enums.Major;
-import com.example.daedongv3_5.domain.recruitment.domain.RecruitmentStatus;
+import com.example.daedongv3_5.domain.recruitment.domain.enums.RecruitmentStatus;
 import lombok.Builder;
 
 import java.util.List;
